src/main/java/com/example/springboottutorial/employee/EmployeeController.java
package com.example.springboottutorial.employee;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import java.net.URI;
import java.util.List;

@RestController
@RequestMapping(path = "/employees")
public class EmployeeController
{
    @Autowired
    private EmployeeService employeeService;

    @GetMapping(path="/", produces = "application/json")
    public List<Employee> getEmployees()
    {
        return employeeService.getEmployeeList();
    }

    @PostMapping(path= "/", consumes = "application/json", produces = "application/json")
    public ResponseEntity<Object> addEmployee(@RequestBody Employee employee) {

        //add resource
        employeeService.addEmployee(employee);

        //Create resource location
        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                .path("/{id}")
                .buildAndExpand(employee.getEmployeeId())
                .toUri();

        //Send location in response
        return ResponseEntity.created(location).build();
    }
}
